# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/groups/{groupId}": {
      "get": {
        "operationId": "GroupsController_getGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get group details (members only)",
        "tags": [
          "Groups"
        ]
      },
      "patch": {
        "operationId": "GroupsController_updateGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update group (OWNER/ADMIN only)",
        "tags": [
          "Groups"
        ]
      },
      "delete": {
        "operationId": "GroupsController_deleteGroup",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete group (OWNER only)",
        "tags": [
          "Groups"
        ]
      }
    },
...
    "/groups/{groupId}/activities/{activityId}/recurrence": {
      "post": {
        "operationId": "EventRecurrenceController_create",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecurrenceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create independently tracked recurring occurrences",
        "tags": [
          "Event Recurrence"
        ]
      }
    },
    "/groups/{groupId}/recurrence-series/{seriesId}/future/{activityId}": {
      "patch": {
        "operationId": "EventRecurrenceController_updateFuture",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seriesId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFutureEventsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update this and future occurrences",
        "tags": [
          "Event Recurrence"
        ]
      }
    },
    "/groups/{groupId}/activities/{activityId}/cancel-occurrence": {
      "post": {
        "operationId": "EventRecurrenceController_cancel",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel one recurring occurrence",
        "tags": [
          "Event Recurrence"
        ]
      }
    },
...
    "/groups/{groupId}/inventory": {
      "get": {
        "operationId": "InventoryController_list",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "holderId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search and filter persistent team inventory",
        "tags": [
          "Team Inventory"
        ]
      },
      "post": {
        "operationId": "InventoryController_create",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a persistent inventory catalogue entry",
        "tags": [
          "Team Inventory"
        ]
      }
    },
    "/groups/{groupId}/inventory/{itemId}": {
      "get": {
        "operationId": "InventoryController_get",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get inventory with availability and custody",
        "tags": [
          "Team Inventory"
        ]
      },
      "patch": {
        "operationId": "InventoryController_update",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInventoryItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team Inventory"
        ]
      },
      "delete": {
        "operationId": "InventoryController_archive",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team Inventory"
        ]
      }
    },
    "/groups/{groupId}/inventory/{itemId}/batches": {
      "post": {
        "operationId": "InventoryController_createBatch",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team Inventory"
        ]
      }
    },
    "/groups/{groupId}/inventory/{itemId}/assets": {
      "post": {
        "operationId": "InventoryController_createAsset",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInventoryAssetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team Inventory"
        ]
      }
    },
    "/groups/{groupId}/inventory/locations/all": {
      "get": {
        "operationId": "InventoryController_listLocations",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team Inventory"
        ]
      }
    },
    "/groups/{groupId}/inventory/locations": {
      "post": {
        "operationId": "InventoryController_createLocation",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStorageLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team Inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "PackPlay API",
    "description": "PackPlay backend API for group equipment management",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Group name",
            "example": "Sunday Football"
          },
          "description": {
            "type": "string",
            "description": "Group description",
            "example": "Weekly football meetup"
          },
          "sportType": {
            "type": "string",
            "description": "Sport type",
            "example": "football"
          }
        }
      },
...
      "CreateRecurrenceDto": {
        "type": "object",
        "properties": {
          "frequency": {
            "type": "string",
            "enum": [
              "DAILY",
              "WEEKLY",
              "CUSTOM"
            ]
          },
          "interval": {
            "type": "number",
            "minimum": 1,
            "maximum": 52,
            "default": 1
          },
          "daysOfWeek": {
            "description": "UTC weekdays, Sunday is 0",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "occurrenceDates": {
            "description": "Explicit ISO timestamps used for CUSTOM recurrence",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timezone": {
            "type": "string",
            "example": "Europe/Warsaw",
            "maxLength": 100
          },
          "until": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "number",
            "minimum": 2,
            "maximum": 366
          }
        },
        "required": [
          "frequency",
          "timezone"
        ]
      },
      "UpdateFutureEventsDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "venueName": {
            "type": "string",
            "maxLength": 200
          },
          "environment": {
            "type": "string",
            "enum": [
              "INDOOR",
              "COVERED_OUTDOOR",
              "OUTDOOR"
            ]
          },
          "surface": {
            "type": "string",
            "maxLength": 100
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "CANCELLED",
              "COMPLETED",
              "ARCHIVED"
            ]
          }
        }
      },
...
      "CreateInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "category": {
            "type": "string",
            "maxLength": 100
          },
          "unit": {
            "type": "string",
            "maxLength": 50,
            "default": "piece"
          },
          "trackingType": {
            "type": "string",
            "enum": [
              "BATCH",
              "ASSET"
            ]
          }
        },
        "required": [
          "name",
          "trackingType"
        ]
      },
      "UpdateInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "category": {
            "type": "string",
            "maxLength": 100
          },
          "unit": {
            "type": "string",
            "maxLength": 50,
            "default": "piece"
          },
          "trackingType": {
            "type": "string",
            "enum": [
              "BATCH",
              "ASSET"
            ]
          }
        }
      },
      "CreateInventoryBatchDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 1
          },
          "condition": {
            "type": "string",
            "enum": [
              "GOOD",
              "NEEDS_ATTENTION",
              "DAMAGED",
              "INCOMPLETE",
              "RETIRED"
            ]
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "quantity"
        ]
      },
      "CreateInventoryAssetDto": {
        "type": "object",
        "properties": {
          "assetTag": {
            "type": "string",
            "maxLength": 100
          },
          "serialNumber": {
            "type": "string",
            "maxLength": 200
          },
          "condition": {
            "type": "string",
            "enum": [
              "GOOD",
              "NEEDS_ATTENTION",
              "DAMAGED",
              "INCOMPLETE",
              "RETIRED"
            ]
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "assetTag"
        ]
      },
      "CreateStorageLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "details": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}